requirements.txt
Django
requests
mysqlclient
django-import-export